Added code to enable STARTTLS for LDAP connections
Added new option, startTLS which defaults to false.  When enabled, the
new startTLS method from ldapts.Client will be invoked before attempting
to bind or query.

The newer version of ldapts changes how it determines if an SSL
connection should be attempted.  Because of this change, passing an
empty tlsOptions when creating the ldapts.Client object is inappropriate
for non-SSL connections.  The setup now resets tlsOptions to be
undefined if the uri doesn't start with "ldaps:", which matches the old
behavior as far as I can determine.

Also bumped version requirement of ldapts to 2.2.0, since that's
required for STARTTLS.